0.8.4 20210901
In working version use of controller object of wsaonepage with propertie
adapted to the displayed component replaced by the use of a MVCFactory
object for eacht type of component that creates a new display controller
for each component to display so we don't need to backup adapt and
restore the properties of the wsaonepage controller object.
Exception is de Input object although we could probably also use e new
copy for each menu item we keep backing up the original,
 filling tha active one from the menu-item and restoring it after use.
Because we work in J4 with new objects with namespaces we also need less
adjustments for paths.
 Cleaned up most code that was needed for the old method.